requirements.txt
pandas>=2.2.2
numpy>=1.26.4
scikit-learn>=1.5.1
mlflow>=2.14.0
pandera>=0.20.3
openpyxl>=3.1.5
requests>=2.32.3
pydantic>=2.9.2
evidently>=0.4.36
matplotlib>=3.9.0
joblib>=1.4.2
fastapi>=0.114.1
uvicorn>=0.30.6
pyyaml>=6.0.2
python-dotenv>=1.0.1
